wip on network spec creation
