+++ crates/orchestrator/src/lib.rs
@@ -37,16 +37,40 @@ where
 
     pub async fn spawn(&self, network_config: NetworkConfig) -> Result<(), OrchestratorError> {
         let global_timeout = network_config.global_settings().network_spawn_timeout();
+        let network_spec = NetworkSpec::from_config(&network_config).await?;
         let r = timeout(
             Duration::from_secs(global_timeout.into()),
-            self.spawn_inner(&NetworkSpec::from_config(&network_config).await?),
+            self.spawn_inner(network_spec),
         )
         .await
         .map_err(|_| OrchestratorError::GlobalTimeOut(global_timeout))?;
         r
     }
 
-    async fn spawn_inner(&self, _network_spec: &NetworkSpec) -> Result<(), OrchestratorError> {
+    async fn spawn_inner(&self, network_spec: NetworkSpec) -> Result<(), OrchestratorError> {
+        // main dirver for spawn the network
+
+        // init provider & create namespace
+
+        // static setup
+
+        // Creta chain-spec for relaychain
+
+        // Create parachain artifacts (chain-spec, wasm, state)
+
+        // Customize relaychain
+
+        // spawn first node of relay-chain and any parachain
+
+        // spawn the rest of the nodes (in batches)
+
+        // add-ons (introspector/tracing/etc)
+
+        // verify nodes (clean metrics cache?)
+
+        // write zombie.json state file
+
+        // return `Network` instance
         Ok(())
     }
 }
